triage-api/requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
numpy
opencv-python-headless
google-generativeai
python-multipart